// tool/model/people/Occupation.ts
export enum Occupation {
  artist = "artist",
  anthropologist = "anthropologist",
  astronaut = "astronaut",
  astronomer = "astronomer",
  astrophysicist = "astrophysicist",
  biologist = "biologist",
  biochemist = "biochemist",
  biophysicist = "biophysicist",
  chemist = "chemist",
  cryptozoologist = "cryptozoologist",
  dentist = "dentist",
  director = "director",
  engineer = "engineer",
  geologist = "geologist",
  geophysicist = "geophysicist",
  geographer = "geographer",
  historian = "historian",
  journalist = "journalist",
  lawyer = "lawyer",
  librarian = "librarian",
  mathematician = "mathematician",
  meteorologist = "meteorologist",
  military = "military",
  musician = "musician",
  neurologist = "neurologist",
  neuroscientist = "neuroscientist",
  oceanographer = "oceanographer",
  parapsychologist = "parapsychologist",
  philosopher = "philosopher",
  painter = "painter",
  pilot = "pilot",
  photographer = "photographer",
  physician = "physician",
  physicist = "physicist",
  politician = "politician",
  producer = "producer",
  psychologist = "psychologist",
  psychiatrist = "psychiatrist",
  sailor = "sailor",
  sociologist = "sociologist",
  softwareEngineer = "softwareEngineer",
  stuntman = "stuntman",
  teacher = "teacher",
  ufologist = "ufologist",
  writer = "writer"
}